+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Release 4.2
+
+ * [close #69] - documentation improvements
+ * [close #70] - property become mandatory when a value or array of value
+ * [close #71] - remove duplication and set methods as deprecated
+
 # Release 4.1
 
  * [close #61] - allow global configuration